removed old server
